python/pyproject.toml
[tool.poetry]
name = "xxx-sqs-consumer"
version = "0.1.0"
description = ""
authors = ["k_kitamura <19430599+tamken@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.28.66"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.10.0"
flake8 = "^6.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
